refactored for community bundle
