+++ examples/simple/m4_test.py
@@ -6,7 +6,7 @@
 '''
 
 import time, struct, board, digitalio as dio
-from circuitpython_nrf24l01.rf24 import RF24
+from circuitpython_nrf24l01 import RF24
 
 # addresses needs to be in a buffer protocol object (bytearray)
 address = b'1Node'
