+++ setup.py
@@ -60,5 +60,5 @@ setup(
     # simple. Or you can use find_packages().
     # TODO: IF LIBRARY FILES ARE A PACKAGE FOLDER,
     #       CHANGE `py_modules=['...']` TO `packages=['...']`
-    packages=['circuitpython_nrf24l01'],
+    py_modules=['circuitpython_nrf24l01'],
 )
